resolved comments on PR
